Seperate async testing to new file.
