+++ tox.ini
@@ -240,7 +240,6 @@ deps =
     framework_flask-flasklatest: flask[async]
     framework_flask-flaskmaster: https://github.com/pallets/werkzeug/archive/main.zip
     framework_flask-flaskmaster: https://github.com/pallets/flask/archive/main.zip#egg=flask[async]
-    framework_graphql: asyncio
     framework_graphql-graphql02: graphql-core<3
     framework_graphql-graphql03: graphql-core<4
     framework_graphql-graphql0201: graphql-core<2.2
